refactor: decompose large modules and add portable quality rules
File size reductions:
- emitter.py: 1369 -> 902 (extracted lang_parsers.py, vendored.py)
- java_generator.py: 747 -> 513 (extracted maven_parser.py)
- runners.py: 625 -> 216 (extracted lang_runners.py)
- compare.py: 496 -> 318 (extracted compare_report.py, DRY fix)

Other fixes:
- Add licenseListVersion to Java SPDX creationInfo
- Update mock paths in tests for refactored modules
- Add portable .windsurf/rules/quality/ folder (14 files)
  with per-language best practices for Python, Rust, Go, Java, C/C++

All 781 tests pass.
